Bug 794127 - hal::AdjustSystemClock should take a 64-bit value. r=cjones
UltraBlame original commit: 6b444917f491c4d4fb5ea9a64a3f63b8860acb34

+++ hal/Hal.h
@@ -242,7 +242,7 @@ void NotifyNetworkChange(const hal::NetworkInformation& aNetworkInfo);
 
 
 
-void AdjustSystemClock(int32_t aDeltaMilliseconds);
+void AdjustSystemClock(int64_t aDeltaMilliseconds);
